now reusing integers

+++ app/src/main/java/db_project/utils/pathFind/dataStructure/Edge.java
@@ -3,9 +3,9 @@ package db_project.utils.pathFind.dataStructure;
 public class Edge {
     private int source;
     private int destination;
-    private double weight;
+    private int weight;
 
-    public Edge(final int src, final int dst, final double weight) {
+    public Edge(final int src, final int dst, final int weight) {
         this.source = src;
         this.destination = dst;
         this.weight = weight;
@@ -27,11 +27,11 @@ public class Edge {
         this.destination = destination;
     }
 
-    public double getWeight() {
+    public int getWeight() {
         return weight;
     }
 
-    public void setWeight(double weight) {
+    public void setWeight(int weight) {
         this.weight = weight;
     }
 }
